lint: lpc1768/msp430: remove arith ops on void ptrs

+++ cpu/cortex-m0_common/thread_arch.c
@@ -66,7 +66,7 @@ char *thread_arch_stack_init(void *(*task_func)(void *),
                              int stack_size)
 {
     uint32_t *stk;
-    stk = (uint32_t *)((uint32_t)stack_start + stack_size);
+    stk = (uint32_t *)((uintptr_t)stack_start + stack_size);
 
     /* marker */
     stk--;
